put files in a files dir

+++ src/index.ts
@@ -5,8 +5,12 @@ import { compressInPlace, sha256HashFile } from "./file";
 
 const args = await getArgs();
 
-await $`rm -rf ${args.clone}`;
-await $`mkdir ${args.clone}`;
+const newFilesDir = path.join(args.clone, "files");
+
+if (args.clean) {
+  await $`rm -rf ${args.clone}`;
+  await $`mkdir ${args.clone}`;
+}
 
 const sourceFiles = new Glob("**/*").scan(args.source);
 
@@ -24,18 +28,18 @@ for await (const sourceFile of sourceFiles) {
   console.log(`Processing ${sourceFile}`);
 
   const sourceFilePath = path.join(args.source, sourceFile);
-  const clonedFilePath = path.join(args.clone, sourceFile);
+  const newFilePath = path.join(newFilesDir, sourceFile);
 
-  await $`mkdir -p ${dirname(clonedFilePath)}`;
-  await $`cp ${sourceFilePath} ${clonedFilePath}`;
-  await compressInPlace(clonedFilePath);
+  await $`mkdir -p ${dirname(newFilePath)}`;
+  await $`cp ${sourceFilePath} ${newFilePath}`;
+  await compressInPlace(newFilePath);
 }
 
 console.log("Generating a manifest file...");
 
-const clonedFiles = new Glob("*/**").scan(args.clone);
+const cloneDir = new Glob("*/**").scan(args.clone);
 
-for await (const clonedFile of clonedFiles) {
+for await (const clonedFile of cloneDir) {
   const clonedFilePath = path.join(args.clone, clonedFile);
 
   console.log(`Calculating meta for ${clonedFilePath}`);
